[DATA-1035] Add option to limit configuring directories (#1843)
This adds a command-line flag to viam-server called `limit-configuring-directories` that can be used to restrict which directories users can configure for storing data on-robot. For now, this is only used by the data management service to restrict data capture to the default data capture directory. This will be used to prevent connections to Try Viam rovers from changing the capture directory.