update gpio_unsupported.go to match gpio.go (#1922)
Thanks to Rand for noticing this problem! In https://github.com/viamrobotics/rdk/pull/1896, I updated the function signature of `gpioInitialize` in the Linux-specific file, but forgot to update the non-Linux version to match.

This PR:
- creates a dummy `gpioPin` struct that contains nothing but implements the GPIOPin interface
- updates `gpioInitialize` to return a map of `gpioPin`s to match the function signature of the Linux version
- removes `gpioGetPin`, which was removed from the Linux version, too

I'm unable to test this myself, and appreciate Jeremy and Rand testing it for me on their Macs.