[RSDK-1862] software PWM for Jetson Orin (#1881)
Tried on the Orin on my desk, using the PWM to blink an LED: works great at lots of duty cycles and low frequencies (up to 5 Hz). When using it to drive a DC motor, things work fine at relatively high motor speeds (e.g., >50% power), but it stutters a lot at low speeds. Similarly, using it to drive a servo motor results in lots of shifting, wiggling movements. Using an oscilloscope, the waveform looks okay for frequencies up to about 50 Hz, but beyond that things become noisy, and by the time you hit 500 Hz, this is garbage. 

We'll aim to nail hardware PWM, and then you won't need to rely on this any more.